+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 # **Change Log**
 All notable changes to this project will be documented in this file.
 
+## **[1.6.2] - 29.01.2024**
+Edited ExceptKey in all class. Minor bug fixes.
+### Added
+ * Added `ExceptKey` method in the `EnumerableDict` class
+
+### Edited
+ * Edited ExceptKey in all class.
+
+<br>
+
 ## **[1.6.1] - 02.11.2023**
-Added abstract classes. Minor bug fixes.
+Added `ToItem` property in all class. Minor bug fixes.
 ### Added
  * Added `ToItem` property in the `EnumerableList`, `EnumerableDict` and `EnumerableItem` class
 ### Fixed